requirements.txt
facebook-sdk
django-sslserver